app/requirements.txt
streamlit==1.23.1
pinecone-client==2.2.2
sentence-transformers==2.2.2
openai==0.27.0
langchain==0.0.188
python-dotenv==1.0.0
typing_extensions==4.6.3
typing-inspect==0.9.0
python-dotenv==1.0.0
Pillow==9.5.0
transformers==4.30.2
boto3==1.26.160
botocore==1.29.160
